requirements.txt
discord
colorama
motor
numpy
pymongo
requests
PyNaCl
mcstatus
aiosqlite
colorama
gspread
aioschedule
yt-dlp
python-dotenv
schedule
psutil
pytz
feedparser